onnx/requirements.txt
fastapi
pillow
onnxruntime
numpy
confluent-kafka
pyyaml
uvicorn